Validation happens in `recover_signer()` (#685)
